add authoriities and annotation

+++ Lab2_8/src/main/java/com/example/skvortsoff/entity/Course.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.io.Serializable;
 import java.math.BigDecimal;
 @Table(name = "course", indexes = {
         @Index(name = "CATEGORY", columnList = "CATEGORY"),
@@ -21,7 +22,7 @@ public class Course {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "NAME", length = 100)
     private String name;
